feat: adding performance data

+++ src/main/java/org/heigvd/dto/workout_dto/data_point_dto/WorkoutPerfDetailsDto.java
@@ -3,15 +3,15 @@ package org.heigvd.dto.workout_dto.data_point_dto;
 public class WorkoutPerfDetailsDto {
 
     private int blocId;
-    private int plannedBPMMin;
-    private int plannedBPMMax;
-    private int actualBPMMean;
+    private double plannedBPMMin;
+    private double plannedBPMMax;
+    private double actualBPMMean;
 
     // CONSTRUCTORS ----------------------
 
     public WorkoutPerfDetailsDto() { }
 
-    public WorkoutPerfDetailsDto(int blocId, int plannedBPMMin, int plannedBPMMax, int actualBPMMean) {
+    public WorkoutPerfDetailsDto(int blocId, double plannedBPMMin, double plannedBPMMax, double actualBPMMean) {
         this.blocId = blocId;
         this.plannedBPMMin = plannedBPMMin;
         this.plannedBPMMax = plannedBPMMax;
@@ -23,14 +23,14 @@ public class WorkoutPerfDetailsDto {
     public int getBlocId() { return blocId; }
     public void setBlocId(int blocId) { this.blocId = blocId; }
 
-    public int getPlannedBPMMin() { return plannedBPMMin; }
-    public void setPlannedBPMMin(int plannedBPMMin) { this.plannedBPMMin = plannedBPMMin; }
+    public double getPlannedBPMMin() { return plannedBPMMin; }
+    public void setPlannedBPMMin(double plannedBPMMin) { this.plannedBPMMin = plannedBPMMin; }
 
-    public int getPlannedBPMMax() { return plannedBPMMax; }
-    public void setPlannedBPMMax(int plannedBPMMax) { this.plannedBPMMax = plannedBPMMax; }
+    public double getPlannedBPMMax() { return plannedBPMMax; }
+    public void setPlannedBPMMax(double plannedBPMMax) { this.plannedBPMMax = plannedBPMMax; }
 
-    public int getActualBPMMean() { return actualBPMMean; }
-    public void setActualBPMMean(int actualBPMMean) { this.actualBPMMean = actualBPMMean; }
+    public double getActualBPMMean() { return actualBPMMean; }
+    public void setActualBPMMean(double actualBPMMean) { this.actualBPMMean = actualBPMMean; }
 
     @Override
     public String toString() {
